refactor: separate api logic scripts
